progress: basic navigation + a GUI

+++ client/app/src/main/java/com/ezntek/beanmaths/components/Component.java
@@ -2,7 +2,8 @@ package com.ezntek.beanmaths.components;
 
 public abstract class Component {
     public ComponentState state;
-    
+
     public abstract void render();
-    public abstract void refresh(int gtState);
+
+    public abstract void refresh(long gtState);
 }
